+++ src/components/classDetail/ClassDetailBtn.tsx
@@ -1,5 +1,6 @@
 import Link from 'next/link';
 import React from 'react';
+import AskButton from '../chatRooms/AskButton';
 
 //Todo :  href chat ID, 받아서 입력할것
 const ClassDetailBtn = ({ classId }: { classId: string }) => {
@@ -11,12 +12,7 @@ const ClassDetailBtn = ({ classId }: { classId: string }) => {
       >
         신청하기
       </Link>
-      <Link
-        href={'/message'}
-        className="flex justify-center items-center rounded-2xl w-20 border-[#5373FF] border-solid border-[1px] h-9 ml-3"
-      >
-        문의하기
-      </Link>
+      <AskButton classId={classId} />
     </div>
   );
 };
